requirements.txt
django
supabase
python-dotenv
dj-database-url
psycopg[binary]
requests